moved Extension to org.citydb.plugin
