+++ citydb-cli/src/main/java/org/citydb/cli/extension/ExportFormatCommand.java
@@ -22,7 +22,7 @@
 package org.citydb.cli.extension;
 
 import org.citydb.cli.command.Command;
-import org.citydb.plugin.extension.Extension;
+import org.citydb.plugin.Extension;
 
 public interface ExportFormatCommand extends Extension, Command {
 }
