+++ citydb-plugin/src/main/java/module-info.java
@@ -5,6 +5,5 @@ module org.citydb.plugin {
     uses org.citydb.plugin.Plugin;
 
     exports org.citydb.plugin;
-    exports org.citydb.plugin.extension;
     exports org.citydb.plugin.metadata;
 }
